fix: profile credentials provider tests
