+++ aws-runtime/aws-config/common/test/aws/sdk/kotlin/runtime/auth/credentials/ProfileCredentialsProviderTest.kt
@@ -9,7 +9,6 @@ import aws.sdk.kotlin.runtime.auth.credentials.internal.credentials
 import aws.sdk.kotlin.runtime.client.AwsClientOption
 import aws.sdk.kotlin.runtime.http.interceptors.businessmetrics.AwsBusinessMetric
 import aws.sdk.kotlin.runtime.http.interceptors.businessmetrics.withBusinessMetric
-import aws.sdk.kotlin.runtime.http.interceptors.businessmetrics.withBusinessMetrics
 import aws.sdk.kotlin.runtime.util.testAttributes
 import aws.smithy.kotlin.runtime.auth.awscredentials.Credentials
 import aws.smithy.kotlin.runtime.auth.awscredentials.copy
@@ -17,10 +16,7 @@ import aws.smithy.kotlin.runtime.collections.attributesOf
 import aws.smithy.kotlin.runtime.httptest.TestConnection
 import aws.smithy.kotlin.runtime.httptest.buildTestConnection
 import aws.smithy.kotlin.runtime.net.Host
-import aws.smithy.kotlin.runtime.time.Instant
 import aws.smithy.kotlin.runtime.util.TestPlatformProvider
-import io.mockk.coEvery
-import io.mockk.mockkStatic
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -375,59 +371,6 @@ class ProfileCredentialsProviderTest {
                 AwsBusinessMetric.Credentials.CREDENTIALS_STS_ASSUME_ROLE,
             ),
         )
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun processBusinessMetrics() = runTest {
-        val testProvider = TestPlatformProvider(
-            env = mapOf(
-                "AWS_CONFIG_FILE" to "config",
-            ),
-            fs = mapOf(
-                "config" to """
-                [default]
-                credential_process = awscreds-custom
-                """.trimIndent(),
-                "awscreds-custom" to "some-process",
-            ),
-        )
-        val testEngine = TestConnection()
-        val provider = ProfileCredentialsProvider(
-            platformProvider = testProvider,
-            httpClient = testEngine,
-        )
-
-        mockkStatic(::executeCommand)
-        coEvery { executeCommand(any(), any(), any(), any(), any()) }.returns(
-            Pair(
-                0,
-                """
-            {
-                "Version": 1,
-                "AccessKeyId": "AKID-Default",
-                "SecretAccessKey": "Default-Secret",
-                "SessionToken": "SessionToken",
-                "Expiration" : "2019-05-29T00:21:43Z"
-            }
-                """.trimIndent(),
-            ),
-        )
-
-        val actual = provider.resolve()
-        val expected = credentials(
-            "AKID-Default",
-            "Default-Secret",
-            "SessionToken",
-            Instant.fromIso8601("2019-05-29T00:21:43Z"),
-            "Process",
-        ).withBusinessMetrics(
-            setOf(
-                AwsBusinessMetric.Credentials.CREDENTIALS_PROFILE_PROCESS,
-                AwsBusinessMetric.Credentials.CREDENTIALS_PROCESS,
-            ),
-        )
-
         assertEquals(expected, actual)
     }
 }
